Refactor composer redux filename
